requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.1.2
django-import-export==4.1.1
django-otp==1.5.4
pillow==10.4.0
sqlparse==0.5.1
tablib==3.5.0
